start welcome comp.
