+++ project/src/app/components/addform/addform.component.html
@@ -40,15 +40,15 @@
           </div>          
           <div class="form-group">
             <label for="CourseStatus">Course Status</label>
-              <select #status>
-                <option value="0">Select</option>
+              <select #status class="form-control">
+                <option value="0">Select One</option>
                 <option *ngFor="let status of statuses" value="{{ status.name }}">{{ status.name }}</option>
               </select>                
           </div>
           <div class="form-group">
             <label for="CourseCategory">Category</label>
                 <select #category class="form-control">
-                  <option value="0">Select</option>
+                  <option value="0">Select One</option>
                   <option *ngFor="let category of categories" value="{{ category.name }}">{{ category.name }}</option>
                 </select>                
           </div>
