+++ project/src/app/components/home/home.component.html
@@ -3,3 +3,4 @@
 <h5 *ngIf="!auth.authenticated()">You are not logged in, please click on the login link above to log in.</h5>
 
 <courses *ngIf="auth.authenticated()"></courses>
+<welcome *ngIf="!auth.authenticated()"></welcome>
