+++ project/src/app/components/welcome/welcome.component.html
@@ -1,0 +1,1 @@
+welcome
